chore: add showIconButton on AccountItem
- Add showIconButton on AccountItem
- Decouple visibleAccounts, hiddenAccounts list to a hook
