+++ packages/wallet-ui/src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './redux';
 export * from './useCurrentNetwork';
 export * from './useCurrentAccount';
 export * from './useScrollTo';
+export * from './useAccountVisibility';
